Fixed bug with coverage, uint16_t --> uint64_t

+++ src/psass.h
@@ -23,8 +23,8 @@
 struct WindowBaseData {
 
     bool snps[2] {0, 0};  // Pair of boolean for sex-specific snps in (pool1, pool2)
-    uint16_t depth[2] {0, 0};  // Pair of int for coverage in (pool1, pool2)
-    uint16_t nucleotides[2][6] {{0, 0, 0, 0, 0, 0}, {0, 0, 0, 0, 0, 0}};
+    uint32_t depth[2] {0, 0};  // Pair of int for coverage in (pool1, pool2)
+    uint32_t nucleotides[2][6] {{0, 0, 0, 0, 0, 0}, {0, 0, 0, 0, 0, 0}};
 };
 
 
@@ -32,8 +32,8 @@ struct WindowBaseData {
 struct Window {
 
     std::deque<WindowBaseData> data;  // Sliding window data as a deque
-    uint16_t snps_in_window[2] {0, 0};  // Pair of int for average snps in current window for (pool1, pool2)
-    uint16_t depth_in_window[2] {0, 0};  // Pair of int for average coverage in current window for (pool1, pool2)
+    uint32_t snps_in_window[2] {0, 0};  // Pair of int for average snps in current window for (pool1, pool2)
+    uint64_t depth_in_window[2] {0, 0};  // Pair of int for average coverage in current window for (pool1, pool2)
     float fst_in_window = 0.0;
 };
